% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratings.R
\docType{data}
\name{ratings}
\alias{ratings}
\title{Ratings of 1M MovieLens data}
\format{a \code{\link{data.frame}} / \code{\link{data.table}} with 1000209 rows and 3 variables:
\describe{
  \item{user}{ID of user, integer}
  \item{item}{ID of item, integer}
  \item{rating}{rating of item by user, integer in range 1:5}
}}
\source{
\url{http://grouplens.org/datasets/movielens/1m/}
}
\usage{
ratings
}
\description{
A dataset contains 1000209 ratings from 6040 users of 3706 movies
}
\keyword{datasets}

