% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_ratings.R
\name{split_ratings}
\alias{split_ratings}
\title{Split ratings data into train, validation and test sets}
\usage{
split_ratings(ratings_table, proportion = c(0.7, 0.15, 0.15))
}
\arguments{
\item{ratings_table}{\code{\link{data.table}} or \code{\link{data.frame}} of ratings. Should contain 3 columns: user (id of user, integer), item (id of item, integer) and rating (rating of item by user, integer or numeric)}

\item{proportion}{(default c(0.7, 0.15, 0.15)) vector of 3 numeric values, which sums to 1 and indicate proportion of ratings to be set to train, validation and test sets}
}
\value{
List of 6 objects:
\itemize{
  \item train_matrix - sparse matrix with ratings for train
  \item valid_matrix - sparse matrix with ratings for validation
  \item test_matrix - sparse matrix with ratings for test
  \item train_matrix_ones - train_matrix with ratings replaced with ones
  \item valid_matrix_ones - valid_matrix with ratings replaced with ones
  \item test_matrix_ones - test_matrix with ratings replaced with ones
}
}
\description{
Split ratings data into train, validation and test sets
}
\details{
Function splits data into train, validation and test sets and transforms it into sparse matrix objects
}
\examples{

mtx <- split_ratings(ratings)

}

