% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd_build.R
\name{svd_build}
\alias{svd_build}
\title{Build SVD Recommender Model}
\usage{
svd_build(mtx)
}
\arguments{
\item{mtx}{list of matrixes for training, validation and testing recommender, returned by \code{\link{split_ratings}} function}
}
\value{
List of 3 objects:
\itemize{
  \item decomposed_matrix - list of matrixes, returned by \code{\link{svd}} decomposition
  \item user_average - vector of average rating of each user
  \item mtx - list of train, validation and test sparse matrixes
}
}
\description{
Build SVD Recommender Model
}
\details{
Function performs SVD decomposition of user-item rating matrix
}
\examples{

mtx <- split_ratings(ratings)
model <- svd_build(mtx)

}

