% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd_rmse.R
\name{svd_rmse}
\alias{svd_rmse}
\title{Calculate RMSE of SVD Approximation model}
\usage{
svd_rmse(model, r = 10, rmse_type = c("train", "validation", "test"))
}
\arguments{
\item{model}{model object, returned by \code{\link{svd_build}} function}

\item{r}{(default 10) number of latent factors to take into account to reconstruct rating matrix and make prediction}

\item{rmse_type}{(default c("train", "validation", "test")) vector, which indicates matrixes for which RMSE should be calculated}
}
\value{
vector of RMSE based on \code{rmse_type}. Errors are returned in next order: "train", "validation", "test"
}
\description{
Calculate RMSE of SVD Approximation model
}
\details{
Function calculate RMSE for SVD Approximation Recommender model for train, validation and test sets
}
\examples{

mtx <- split_ratings(ratings)
model <- svd_build(mtx)
svd_rmse(model, r = 10, rmse_type = c("validation"))
svd_rmse(model, r = 5, rmse_type = c("train", "validation"))

}

