% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd_tune.R
\name{svd_tune}
\alias{svd_tune}
\title{Tune SVD Recommender Model}
\usage{
svd_tune(model, r = 2:50, color_pallet = "Accent")
}
\arguments{
\item{model}{model object, returned by \code{\link{svd_build}} function}

\item{r}{(default 2:50) vector of number of latent factors to take into account while reconstructing rating matrix. From these values will be selected one with the smallest RMSE on validation set}

\item{color_pallet}{(default "Accent") name of palett to use to color graphs. See \code{\link{brewer.pal}}}
}
\value{
List of 3 objects:
\itemize{
  \item r_best - best value of r which was selected based on validation error
  \item all_rmse - \code{\link{data.frame}} with train and validation errors for each r
  \item train_vs_valid - ggplot graph of train and validation errors for different r
}
}
\description{
Tune SVD Recommender Model
}
\details{
Function selects the best value for number of latent factors to minimize validation error
}
\examples{

mtx <- split_ratings(ratings)
model <- svd_build(mtx)
svd_tune(model, r = 2:5)

}

