% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_ratings.R
\name{visualize_ratings}
\alias{visualize_ratings}
\title{Visualize statistics for all ratings, users' ratings and items' ratings}
\usage{
visualize_ratings(ratings_table, color_palett = "Accent")
}
\arguments{
\item{ratings_table}{\code{\link{data.table}} or \code{\link{data.frame}} of ratings. Should contain 3 columns: user (id of user, integer), item (id of item, integer) and rating (rating of item by user, integer or numeric)}

\item{color_palett}{(default "Accent") name of palett to use to color graphs. See \code{\link{brewer.pal}}}
}
\value{
5 graphs combined into one plot:
\itemize{
  \item count of different ratings
  \item histogram of users' average ratings
  \item histogram of items' average ratings
  \item histogram of number of rated items by user
  \item histogram of number of scores items have
}
}
\description{
Visualize statistics for all ratings, users' ratings and items' ratings
}
\details{
Function combines 5 ggplot graphs into 1 plot and returns it
}
\examples{

visualize_ratings(ratings_table = ratings)
visualize_ratings(ratings_table = ratings, color_palett = "Dark2")

}

